+++ changelog.md
@@ -10,6 +10,14 @@ Group Members:
 [MATE #2] Jeremy Lee
 [MATE #3] Michaela Yoon
 
+## 0.0.15 - 2018-03-27
+### Added by Michaela
+- application/controllers/Customization.php - section for mix/match accessories
+- application/controllers/Roles.php - controller for roles
+
+### Modified by Michaela
+- application/config/constants.php - added constants for user roles (guest, user, admin)
+- application/views/template.php - shows section linking to customization, edited role section to link to Role controller
 
 ## 0.0.14 - 2018-02-11
 ### Modified by Alfred Swinton